refactor(context-menu): H.9 — extract useContextMenuRating hook
