+++ src/components/ContextMenu.tsx
@@ -1,5 +1,5 @@
 import { getPlaylists, getPlaylist, updatePlaylist } from '../api/subsonicPlaylists';
-import { star, unstar, setRating } from '../api/subsonicStarRating';
+import { star, unstar } from '../api/subsonicStarRating';
 import { getAlbum } from '../api/subsonicLibrary';
 import { getArtist } from '../api/subsonicArtists';
 import type { SubsonicAlbum, SubsonicArtist, SubsonicPlaylist } from '../api/subsonicTypes';
@@ -46,6 +46,7 @@ import {
   startRadio as startRadioAction,
 } from '../utils/contextMenuActions';
 import { useContextMenuKeyboardNav } from '../hooks/useContextMenuKeyboardNav';
+import { useContextMenuRating } from '../hooks/useContextMenuRating';
 
 export { AddToPlaylistSubmenu };
 
@@ -73,7 +74,6 @@ export default function ContextMenu() {
     }))
   );
   const auth = useAuthStore();
-  const setEntityRatingSupport = useAuthStore(s => s.setEntityRatingSupport);
   const entityRatingSupport =
     auth.activeServerId ? auth.entityRatingSupportByServer[auth.activeServerId] ?? 'unknown' : 'unknown';
   const audiomuseNavidromeEnabled = !!(auth.activeServerId && auth.audiomuseNavidromeByServer[auth.activeServerId]);
@@ -135,98 +135,8 @@ export default function ContextMenu() {
   const isStarred = (id: string, itemStarred?: string) =>
     id in starredOverrides ? starredOverrides[id] : !!itemStarred;
 
-  const applySongRating = useCallback((songId: string, rating: number) => {
-    setUserRatingOverride(songId, rating);
-    setRating(songId, rating).catch(() => {});
-  }, [setUserRatingOverride]);
-
-  const applyAlbumRating = useCallback((album: SubsonicAlbum, rating: number) => {
-    setUserRatingOverride(album.id, rating);
-    if (entityRatingSupport !== 'full') return;
-    setRating(album.id, rating).catch(err => {
-      if (auth.activeServerId) setEntityRatingSupport(auth.activeServerId, 'track_only');
-      showToast(
-        typeof err === 'string' ? err : err instanceof Error ? err.message : t('entityRating.saveFailed'),
-        4500,
-        'error',
-      );
-    });
-  }, [setUserRatingOverride, entityRatingSupport, auth.activeServerId, setEntityRatingSupport, t]);
-
-  const applyArtistRating = useCallback((artist: SubsonicArtist, rating: number) => {
-    setUserRatingOverride(artist.id, rating);
-    if (entityRatingSupport !== 'full') return;
-    setRating(artist.id, rating).catch(err => {
-      if (auth.activeServerId) setEntityRatingSupport(auth.activeServerId, 'track_only');
-      showToast(
-        typeof err === 'string' ? err : err instanceof Error ? err.message : t('entityRating.saveFailed'),
-        4500,
-        'error',
-      );
-    });
-  }, [setUserRatingOverride, entityRatingSupport, auth.activeServerId, setEntityRatingSupport, t]);
-
-  const getRatingValueByKind = useCallback((kind: 'song' | 'album' | 'artist', id: string): number => {
-    if (kind === 'song' && (type === 'song' || type === 'album-song' || type === 'queue-item')) {
-      const song = item as Track;
-      if (song.id === id) return userRatingOverrides[id] ?? song.userRating ?? 0;
-    }
-    if (kind === 'album' && type === 'album') {
-      const album = item as SubsonicAlbum;
-      if (album.id === id) return userRatingOverrides[id] ?? album.userRating ?? 0;
-    }
-    if (kind === 'album' && type === 'multi-album') {
-      const albums = item as SubsonicAlbum[];
-      const compositeId = [...albums.map(a => a.id)].sort().join('\x1e');
-      if (id !== compositeId) return userRatingOverrides[id] ?? 0;
-      if (albums.length === 0) return 0;
-      const vals = albums.map(a => userRatingOverrides[a.id] ?? a.userRating ?? 0);
-      const first = vals[0];
-      return vals.every(v => v === first) ? first : 0;
-    }
-    if (kind === 'artist' && type === 'artist') {
-      const artist = item as SubsonicArtist;
-      if (artist.id === id) return userRatingOverrides[id] ?? artist.userRating ?? 0;
-    }
-    if (kind === 'artist' && type === 'multi-artist') {
-      const artists = item as SubsonicArtist[];
-      const compositeId = [...artists.map(a => a.id)].sort().join('\x1e');
-      if (id !== compositeId) return userRatingOverrides[id] ?? 0;
-      if (artists.length === 0) return 0;
-      const vals = artists.map(a => userRatingOverrides[a.id] ?? a.userRating ?? 0);
-      const first = vals[0];
-      return vals.every(v => v === first) ? first : 0;
-    }
-    return userRatingOverrides[id] ?? 0;
-  }, [type, item, userRatingOverrides]);
-
-  const commitRatingByKind = useCallback((kind: 'song' | 'album' | 'artist', id: string, rating: number) => {
-    if (kind === 'song') {
-      applySongRating(id, rating);
-      return;
-    }
-    if (kind === 'album' && type === 'album') {
-      applyAlbumRating(item as SubsonicAlbum, rating);
-      return;
-    }
-    if (kind === 'album' && type === 'multi-album') {
-      const albums = item as SubsonicAlbum[];
-      const compositeId = [...albums.map(a => a.id)].sort().join('\x1e');
-      if (id !== compositeId) return;
-      for (const a of albums) applyAlbumRating(a, rating);
-      return;
-    }
-    if (kind === 'artist' && type === 'artist') {
-      applyArtistRating(item as SubsonicArtist, rating);
-      return;
-    }
-    if (kind === 'artist' && type === 'multi-artist') {
-      const artists = item as SubsonicArtist[];
-      const compositeId = [...artists.map(a => a.id)].sort().join('\x1e');
-      if (id !== compositeId) return;
-      for (const a of artists) applyArtistRating(a, rating);
-    }
-  }, [applySongRating, applyAlbumRating, applyArtistRating, type, item]);
+  const { applySongRating, applyAlbumRating, applyArtistRating, getRatingValueByKind, commitRatingByKind } =
+    useContextMenuRating({ type, item, userRatingOverrides, setUserRatingOverride, entityRatingSupport, t });
 
   const { onMenuKeyDown } = useContextMenuKeyboardNav({
     menuRef,
